checkpoint: rename files
