+++ src/layup/utilities/file_access_utils.py
@@ -2,7 +2,7 @@ import os
 import sys
 
 
-def FindFileOrExit(arg_fn, argname):
+def find_file_or_exit(arg_fn, argname):
     """Checks to see if a file given by a filename exists. If it doesn't,
     this fails gracefully and exits to the command line.
 
@@ -27,7 +27,7 @@ def FindFileOrExit(arg_fn, argname):
         sys.exit("ERROR: filename {} supplied for {} argument does not exist.".format(arg_fn, argname))
 
 
-def FindDirectoryOrExit(arg_fn, argname):
+def find_directory_or_exit(arg_fn, argname):
     """Checks to see if a directory given by a filepath exists. If it doesn't,
     this fails gracefully and exits to the command line.
